+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [16.0.0] - 2019-09-25
+
+### Added
+
+- Started using TypeScript, now shipping types with the package dist.
+
 ### Changed
 
 - The `.chq-tsl--op-act` class over to `.chq-tsl--op[aria-current='true']` to be more consistent and more accessible.
@@ -1046,7 +1052,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Initial checkin.
 - The `Warning` component.
 
-[unreleased]: https://github.com/CultureHQ/components/compare/v15.0.1...HEAD
+[unreleased]: https://github.com/CultureHQ/components/compare/v16.0.0...HEAD
+[16.0.0]: https://github.com/CultureHQ/components/compare/v15.0.1...v16.0.0
 [15.0.1]: https://github.com/CultureHQ/components/compare/v15.0.0...v15.0.1
 [15.0.0]: https://github.com/CultureHQ/components/compare/v14.0.0...v15.0.0
 [14.0.0]: https://github.com/CultureHQ/components/compare/v13.4.0...v14.0.0